refactor(usage): extract antigravity and minimax from usage.ts (#3594)
Extract 1048 lines from the 3256-line usage.ts:
- usage/antigravity.ts (687 lines) — Antigravity/Gemini CLI usage
- usage/minimax.ts (361 lines) — MiniMax usage functions

usage.ts → usage/index.ts (3256 → 2223 lines).
Pure extraction, no behavioral changes. Part of #3594.
